% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.r
\name{bbg_equity_BS}
\alias{bbg_equity_BS}
\title{Balance sheet historical data from Bloomberg}
\usage{
bbg_equity_BS(tickers = "NEM US Equity", start = "2018-01-01",
  end = "2018-06-30", ...)
}
\arguments{
\item{tickers}{A chatacter vector. Specifies the Bloomberg equity tickers to query data for.
Defaults to 'NEM US Equity', the Bloomberg equity ticker for the Newmont Minning Corporation.}

\item{start}{A scalar character vector. Specifies the starting date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-01-01'.}

\item{end}{A scalar character vector. Specifies the end date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-06-30'.}

\item{...}{Optional parameters to pass to the \code{\link[Rblpapi]{bdh}} function from the \code{Rblpapi} package used
for the query (\code{options} parameter).}
}
\value{
An S4 object of class \code{\linkS4class{EquityBS}} with slots:
  \itemize{
    \item{\code{tickers}: a character vector. Equity Bloomberg tickers for which data has been found.}
    \item{\code{fields}: a tibble. Columns include:
      \itemize{
        \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
        \item{\code{section}: balance sheet sections ('asset', 'liabilities') for which data has been found.}
        \item{\code{subsection}: balance sheet subsections ('asset': 'current', 'long term'; 'liabilities':
          'debt', 'equity') for which data has been found.}
        \item{\code{name}: balance sheet Bloomberg data field names for which data has been found.}
      }
    }
    \item{\code{data}: a tibble. Columns include:
      \itemize{
        \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
        \item{\code{section}: balance sheet sections ('asset', 'liabilities') for which data has been found.}
        \item{\code{subsection}: balance sheet subsections ('asset': 'current', 'long term'; 'liabilities':
          'debt', 'equity') for which data has been found.}
        \item{\code{name}: balance sheet Bloomberg data field names for which data has been found.}
        \item{\code{date}: observation date.}
        \item{\code{value}: corresponding observation.}
      }
    }
    \item{\code{call}: a scalar character vector showing the original call to the constructor.}
  }
}
\description{
Provided with a set of Bloomberg equity tickers and a time period,
  queries Bloomberg for the corresponding balance sheet historical data.
}
\examples{
\dontrun{bbg_equity_BS()}

}
\seealso{
The \code{\link[bbgsymbols]{fields}} dataset in the \code{bbgsymbols} package for details on the Bloomnerg fields used here.
}
