% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{BBG_equity_info}
\alias{BBG_equity_info}
\title{Retrieves corporate qualitative information from Bloomberg.}
\usage{
BBG_equity_info(tickers, ...)
}
\arguments{
\item{tickers}{a chatacter vector. Specifies the equity Bloomberg tickers to query
data for.}

\item{...}{optional parameters to pass to the \link[Rblpapi]{bdp} function from the
\href{http://dirk.eddelbuettel.com/code/rblpapi.html}{\pkg{Rblpapi}} package used
  for the query (\code{options} parameter).}
}
\value{
An S4 object of class \linkS4class{EquityInfo}.
}
\description{
Provided with a set of equity Bloomberg tickers retrieves qualitative
  information on the corresponding corporations.
}
\examples{
\dontrun{
    BBG_equity_info(tickers = c("BP/ LN Equity", "WEIR LN Equity"))
  }


}
\seealso{
\itemize{

    \item{The \link[BBGsymbols]{fields} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
     (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite) for details
     on the Bloomnerg fields used here.}

    \item{Helper datasets in the
    \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} and
    \href{https://github.com/bautheac/GICS/}{\pkg{GICS}} packages
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}


  }
}
