% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{bbg_fund_market}
\alias{bbg_fund_market}
\title{Fund historical market data from Bloomberg}
\usage{
bbg_fund_market(tickers = "GLD US Equity", start = "2018-01-01",
  end = "2018-06-30", verbose = TRUE, ...)
}
\arguments{
\item{tickers}{A chatacter vector. Specifies the Bloomberg fund tickers to query data for.
Defaults to 'GLD US Equity', the Bloomberg fund ticker for the 'SPRD gold shares' ETF, the commodiy ETF
with the highest market capitalization at the time of writing (August, 2018).}

\item{start}{A scalar character vector. Specifies the starting date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-01-01'.}

\item{end}{A scalar character vector. Specifies the end date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-06-30'.}

\item{verbose}{A logical scalar vector. Should progression messages be printed? Defaults to TRUE.}

\item{...}{Optional parameters to pass to the \code{\link[Rblpapi]{bdh}} function from the \code{Rblpapi} package used
for the query (\code{options} parameter).}
}
\value{
An S4 object of class \code{\linkS4class{FundMarket}} with slots:
  \itemize{
    \item{\code{tickers}: a tibble. Columns include:
      \itemize{
        \item{\code{ticker}: fund Bloomberg tickers for which data has been found.}
        \item{\code{field}: fund Bloomberg information data fields. Descriptive elements on the fund identified.}
        \item{\code{name}: corresponding information data field names.}
        \item{\code{value}: corresponding information data field value.}#'
      }
    }
    \item{\code{fields}: a tibble. Columns include:
      \itemize{
        \item{\code{ticker}: fund Bloomberg tickers for which data has been found.}
        \item{\code{field}: fund Bloomberg market data fields for which data has been found.}
      }
    }
    \item{\code{data}: a tibble. Columns include:
      \itemize{
        \item{\code{ticker}: fund Bloomberg tickers for which data has been found.}
        \item{\code{field}: fund Bloomberg market data fields for which data has been found.}
        \item{\code{date}: observation date.}
        \item{\code{value}: corresponding value.}
      }
    }
    \item{\code{call}: a scalar character vector showing the original call to the constructor.}
  }
}
\description{
Provided with a set of Bloomberg fund tickers and a time period,
  queries Bloomberg for the corresponding fund historical market data.
}
\examples{
\dontrun{bbg_fund_market()}

}
\seealso{
The \code{\link[bbgsymbols]{fields}} dataset in the \code{bbgsymbols} package for details on the Bloomnerg fields used here.
}
