% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{bbg_futures_aggregate}
\alias{bbg_futures_aggregate}
\title{Futures aggregate historical data from Bloomberg}
\usage{
bbg_futures_aggregate(active_contract_tickers = "C A Comdty",
  start = "2018-01-01", end = "2018-06-30", verbose = TRUE, ...)
}
\arguments{
\item{active_contract_tickers}{A chatacter vector. Specifies the Bloomberg futures active contract tickers to query data for.
Defaults to 'C A Comdty', the Bloomberg active contract ticker for the XCBT corn futures series.}

\item{start}{A scalar character vector. Specifies the starting date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-01-01'.}

\item{end}{A scalar character vector. Specifies the end date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-06-30'.}

\item{verbose}{A logical scalar vector. Should progression messages be printed? Defaults to TRUE.}

\item{...}{Optional parameters to pass to the \code{\link[Rblpapi]{bdh}} function from the \code{Rblpapi} package used
for the query (\code{options} parameter).}
}
\value{
An S4 object of class \code{\linkS4class{FuturesAggregate}} with slots:
  \itemize{
    \item{\code{tickers}: a character vector. Active contract Bloomberg tickers for which data has been found.}
    \item{\code{fields}: a tibble. Columns include:
      \itemize{
        \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
        \item{\code{field}: futures Bloomberg aggregate data fields for which data has been found.}
      }
    }
    \item{\code{data}: a tibble. Columns inlude:
      \itemize{
        \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
        \item{\code{field}: Bloomberg aggregate data fields for which data has been found.}
        \item{\code{date}: observation date.}
        \item{\code{value}: corresponding value.}
      }
    }
    \item{\code{call}: a scalar character vector showing the original call to the constructor.}
  }
}
\description{
Provided with a set of Bloomberg futures active contract tickers and a time period,
  queries Bloomberg for the corresponding futures historical aggregate data. For each individual field,
  futures aggregate data represents the aggregation of the corresponding field values over all the
  corresponding term structure contracts.
}
\examples{
\dontrun{bbg_futures_aggregate()}

}
\seealso{
The \code{fields} dataset in the \code{bbgsymbols} package for details on the Bloomnerg fields used here.
}
