% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{BBG_futures_CFTC}
\alias{BBG_futures_CFTC}
\title{Retrieves futures CFTC reports historical data from Bloomberg.}
\usage{
BBG_futures_CFTC(active_contract_tickers, start, end, verbose = T, ...)
}
\arguments{
\item{active_contract_tickers}{a chatacter vector. Specifies the futures active
contract Bloomberg tickers to query data for.}

\item{start}{a scalar character vector. Specifies the starting date for the query
in the following format: 'yyyy-mm-dd'.}

\item{end}{a scalar character vector. Specifies the end date for the query in the
following format: 'yyyy-mm-dd'.}

\item{verbose}{a logical scalar vector. Should progression messages be printed?
Defaults to TRUE.}

\item{...}{optional parameters to pass to the \link[Rblpapi]{bdh} function from the
\href{http://dirk.eddelbuettel.com/code/rblpapi.html}{\pkg{Rblpapi}} package used
  for the query (\code{options} parameter).}
}
\value{
An S4 object of class \linkS4class{FuturesCFTC}.
}
\description{
Provided with a set of futures active contract Bloomberg tickers and a
  time period, queries Bloomberg for the corresponding futures CFTC reports
  historical data.
}
\examples{
\dontrun{

    BBG_futures_CFTC(active_contract_tickers = c("W A Comdty", "KWA Comdty"),
      start = "2000-01-01", end = as.character(Sys.Date()))

  }


}
\seealso{
\itemize{

    \item{"GFUT <GO>" on a Bloomberg terminal.}

    \item{The \link[BBGsymbols]{tickers_CFTC} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite)
    for details on the Bloomnerg position tickers used here.}

    \item{Helper datasets in the
    \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} package
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}

  }
}
