% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{BBG_futures_info}
\alias{BBG_futures_info}
\title{Futures series qualitative information from Bloomberg.}
\usage{
BBG_futures_info(active_contract_tickers, ...)
}
\arguments{
\item{active_contract_tickers}{A chatacter vector. Specifies the futures active contract Bloomberg tickers to query data for.}

\item{...}{Optional parameters to pass to the \link[Rblpapi]{bdp} function from the
\href{http://dirk.eddelbuettel.com/code/rblpapi.html}{\pkg{Rblpapi}} package used
  for the query (\code{options} parameter).}
}
\value{
An S4 object of class \linkS4class{FuturesInfo}.
}
\description{
Provided with a set of futures active contract Bloomberg tickers retrieves qualitative information
  on the corresponding futures series.
}
\examples{
\dontrun{
    BBG_futures_info(active_contract_tickers = c("W A Comdty", "KWA Comdty"))
  }

}
\seealso{
\itemize{
    \item{The \link[BBGsymbols]{fields} dataset in the \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
     (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite) for details on the Bloomnerg fields used here.}
    \item{Helper datasets in the \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} and
    \href{https://github.com/bautheac/GICS/}{\pkg{GICS}} packages
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}
  }
}
