% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_support.r
\name{bbg_pull_historical}
\alias{bbg_pull_historical}
\title{Pull historical data from Bloomberg}
\usage{
bbg_pull_historical(tickers = "C A Comdty", fields = "PX_LAST",
  start = "2018-01-01", end = "2018-06-30", ...)
}
\arguments{
\item{tickers}{A character vector. Specifies the Bloomberg tickers for the query.
Defaults to 'C A Comdty', the Bloomberg active contract ticker for the XCBT corn futures series.}

\item{fields}{A character vector. Specifies the Bloomberg fields for the query.
Defaults to 'PX_LAST', the Bloomberg close price symbol for the corresponding ticker.}

\item{start}{A scalar character vector. Specifies the starting date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-01-01'.}

\item{end}{A scalar character vector. Specifies the end date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-06-30'.}

\item{...}{Optional parameters to pass to \code{\link[Rblpapi]{bdh}} function from the \code{Rblpapi} package used
for the query (\code{options} parameter).}
}
\value{
A dataframe (if only one ticker supplied in \code{tickers}) or list of dataframes (if more than one tickers supplied
  in \code{tickers}) named after the tickers supplied in \code{tickers}
}
\description{
Given a set of Bloomberg tickers and field symbols, pulls the corresponding data from Bloomberg for the period between
  the dates specified in the \code{start} and \code{end} parameters.
}
\details{
Write more Bloomberg stuff here.
}
\examples{
\dontrun{bbg_pull_historical()}

}
