% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{BBGFuturesHistorical-class}
\alias{BBGFuturesHistorical-class}
\title{S4 class for Bloomberg futures historical data objects}
\description{
S4 class for Bloomberg futures historical data objects
}
