% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{create_futures_spot}
\alias{create_futures_spot}
\title{Creates a FuturesSpot object from futures spot data.}
\usage{
create_futures_spot(data)
}
\arguments{
\item{data}{a dataframe containing futures spot data. Columns should include:

  \itemize{
    \item{ticker: futures active contract Bloomberg tickers.}
    \item{field: field Bloomberg symbols.
      See \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} for a list
      of accepted symbols.}
    \item{date: observation date (YYYY-MM-DD).}
    \item{value: observation value.}
  }}
}
\value{
An S4 object of class \linkS4class{FuturesSpot}.
}
\description{
Futures spot data refers to data on the corresponding cash market.
  Should spot data be retrieved from other data source(s), This function creates
  a FuturesSpot data object out of input data retrieved from any data source.
  The returned FuturesSpot object facilitates storage in a
  \href{https://bautheac.github.io/storethat/}{\pkg{storethat}} database.
}
