% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{dataset,EquityBS-method}
\alias{dataset,EquityBS-method}
\title{'dataset' slot accessor method for S4 objects of class \linkS4class{EquityBS}.}
\usage{
\S4method{dataset}{EquityBS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityBS}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{section}: balance sheet sections ('asset', 'liabilities') for which data has been found.}
    \item{\code{subsection}: balance sheet subsections ('asset': 'current', 'long term'; 'liabilities':
      'debt', 'equity') for which data has been found.}
    \item{\code{name}: balance sheet Bloomberg data field names for which data has been found.}
    \item{\code{date}: observation date.}
    \item{\code{value}: corresponding observation.}
  }
}
\description{
'dataset' slot accessor method for S4 objects of class \linkS4class{EquityBS}.
}
