% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{dataset,EquityRatios-method}
\alias{dataset,EquityRatios-method}
\title{'dataset' slot accessor method for S4 objects of class \linkS4class{EquityRatios}.}
\usage{
\S4method{dataset}{EquityRatios}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityRatios}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{type}: financial ratio types ('profitability', 'margin', 'asset turnover',
      'short term liquidity', 'long term solvency') for which data has been found.}
    \item{\code{name}: financial ratio Bloomberg data field names for which data has been found.}
    \item{\code{date}: observation date.}
    \item{\code{value}: corresponding observation.}
  }
}
\description{
'dataset' slot accessor method for S4 objects of class \linkS4class{EquityRatios}.
}
