% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{dataset,FuturesAggregate-method}
\alias{dataset,FuturesAggregate-method}
\title{'dataset' slot accessor method for S4 objects of class \linkS4class{FuturesAggregate}.}
\usage{
\S4method{dataset}{FuturesAggregate}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesAggregate}.}
}
\value{
A tibble. Columns include:
   \itemize{
     \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
     \item{\code{field}: Bloomberg aggregate data fields for which data has been found.}
     \item{\code{date}: observation date.}
     \item{\code{value}: corresponding value.}
   }
}
\description{
'dataset' slot accessor method for S4 objects of class \linkS4class{FuturesAggregate}.
}
