% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{dataset,FuturesCFTC-method}
\alias{dataset,FuturesCFTC-method}
\title{'dataset' slot accessor method for S4 objects of class \linkS4class{FuturesCFTC}.}
\usage{
\S4method{dataset}{FuturesCFTC}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesCFTC}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
    \item{\code{format}: report formats for which data has been found.}
    \item{\code{underlying}: underlying instruments for which data has been found..}
    \item{\code{unit}: counting units for which data has been found.}
    \item{\code{participant}: trader classifications for which data has been found.}
    \item{\code{position}: trader positions for which data has been found.}
    \item{\code{date}: observation date.}
    \item{\code{value}: observed value}
  }
}
\description{
'dataset' slot accessor method for S4 objects of class \linkS4class{FuturesCFTC}.
}
