% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{dataset,FuturesTS-method}
\alias{dataset,FuturesTS-method}
\title{'dataset' slot accessor method for S4 objects of class \linkS4class{FuturesTS}.}
\usage{
\S4method{dataset}{FuturesTS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesTS}.}
}
\value{
A tibble. Columns include:
   \itemize{
     \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
     \item{\code{TS position}: futures chain term structure positions for which data has been found.}
     \item{\code{field}: futures Bloomberg market data fields for which data has been found.}
     \item{\code{date}: observation date.}
     \item{\code{value}: corresponding value.}
   }
}
\description{
'dataset' slot accessor method for S4 objects of class \linkS4class{FuturesTS}.
}
