% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\name{dataset}
\alias{dataset}
\title{Generic method for accessing the 'dataset' slot of various S4 objects from the \code{pullit} package.}
\usage{
dataset(object)
}
\arguments{
\item{object}{an S4 object with a 'dataset' slot.}
}
\description{
Generic method for accessing the 'dataset' slot of various S4 objects from the \code{pullit} package.
}
