% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{fields,EquityIS-method}
\alias{fields,EquityIS-method}
\title{'fields' slot accessor method for S4 objects of class \linkS4class{EquityIS}.}
\usage{
\S4method{fields}{EquityIS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityIS}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{name}: income statement Bloomberg field names for which data has been found.}
  }
}
\description{
'fields' slot accessor method for S4 objects of class \linkS4class{EquityIS}.
}
