% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{fields,FuturesAggregate-method}
\alias{fields,FuturesAggregate-method}
\title{'fields' slot accessor method for S4 objects of class \linkS4class{FuturesAggregate}.}
\usage{
\S4method{fields}{FuturesAggregate}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesAggregate}.}
}
\value{
A tibble. Columns include:
   \itemize{
     \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
     \item{\code{field}: futures Bloomberg aggregate data fields for which data has been found.}
   }
}
\description{
'fields' slot accessor method for S4 objects of class \linkS4class{FuturesAggregate}.
}
