% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{fields}
\alias{fields}
\title{Generic method for accessing the 'fields' slot of various S4 objects from the \code{pullit} package.}
\usage{
fields(object)
}
\arguments{
\item{object}{an S4 object with a 'fields' slot.}
}
\description{
Generic method for accessing the 'fields' slot of various S4 objects from the \code{pullit} package.
}
