% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_support.r
\name{futures_ticker}
\alias{futures_ticker}
\title{Bloomberg futures term structure ticker}
\usage{
futures_ticker(active_contract_ticker = "C A Comdty", TS_position = 1L,
  roll_type = "A", roll_days = 0L, roll_months = 0L,
  roll_adjustment = "N")
}
\arguments{
\item{active_contract_ticker}{A scalar chatacter vector. Specifies the Bloomberg futures active contract ticker to use for term structure ticker construction.
Defaults to 'C A Comdty', the Bloomberg active contract ticker for the XCBT corn futures series.}

\item{TS_position}{A scalar integer vector. Specifies the term structure position desired. Defaults to 1: front nearby contract for the corresponding futures series.}

\item{roll_type}{A scalar chatacter vector. Specifies roll type to use for term structure ticker construction. Must be one of 'A', 'B', 'D', 'F', 'N', 'O' or 'R'.
Defaults to 'A' or 'With active future': rolls to the most actively traded contract in the futures series.}

\item{roll_days}{A scalar integer vector. Specifies the day the roll should be done. Refers to the day of the month (\code{roll_type} = 'F') or
the number of days before a reference date (\code{roll_type} = 'D', \code{roll_type} = 'N', \code{roll_type} = 'O', \code{roll_type} = 'R'). Works in tandem with `roll_months` below.
Defaults to 0.}

\item{roll_months}{A scalar integer vector. Specifies the month the roll should be done. Refers to the number of months before a
reference date (\code{roll_type} = 'D', \code{roll_type} = 'N', \code{roll_type} = 'O', \code{roll_type} = 'R'). Works in tandem with `roll_days` above.
Defaults to 0.}

\item{roll_adjustment}{A scalar chatacter vector. Specifies roll adjustment method to use for term structure ticker construction.
Must be one of 'D', 'N', 'R', or 'W'. Defaults to 'N' or 'None'.}
}
\value{
A scalar character vector containing the corresponding term structure ticker.
}
\description{
Provided with a futures active contract ticker, a term structure position and a set of roll parameters, constructs the
  corresponding futures term structure Bloomberg ticker according Bloomberg futures term structure ticker construction method.
}
\details{
See 'DOCS #2072138 <GO>' on a Bloomberg terminal to learn more about the Bloomberg rolling conventions.
}
\examples{
futures_ticker(active_contract_ticker = "C A Comdty", TS_position = 5L,
  roll_type = "A", roll_days = 0L, roll_months = 0L, roll_adjustment = "N")

}
\seealso{
\code{\link{bbg_futures_TS}}
}
