% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_call,FuturesTS-method}
\alias{get_call,FuturesTS-method}
\title{'call' slot accessor method for S4 objects of class \linkS4class{FuturesTS}.}
\usage{
\S4method{get_call}{FuturesTS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesTS}.}
}
\value{
A scalar character vector showing the original call to the constructor function,
}
\description{
'call' slot accessor method for S4 objects of class \linkS4class{FuturesTS}.
}
