% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.R
\docType{methods}
\name{get_call}
\alias{get_call}
\alias{get_call,DataInfo-method}
\alias{get_call,DataInfo}
\alias{get_call,DataHistorical-method}
\alias{get_call,DataHistorical}
\title{Accesses the \code{call} slot of various S4 objects output of
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}.}
\usage{
get_call(object)

\S4method{get_call}{DataInfo}(object)

\S4method{get_call}{DataHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataInfo} or
\linkS4class{DataHistorical} or childs.}
}
\value{
The original call to the constructor function.
}
\description{
Access method for the \code{call} slot of various S4
  objects from the \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}
  package.
}
