% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_data,EquityCF-method}
\alias{get_data,EquityCF-method}
\title{'data' slot accessor method for S4 objects of class \linkS4class{EquityCF}.}
\usage{
\S4method{get_data}{EquityCF}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityCF}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{section}: cash flow statement sections ('operating', 'financing', 'investing',
      'total') for which data has been found.}
    \item{\code{name}: cash flow statement Bloomberg field names for which data has been found.}
    \item{\code{date}: observation date.}
    \item{\code{value}: corresponding observation.}
  }
}
\description{
'data' slot accessor method for S4 objects of class \linkS4class{EquityCF}.
}
