% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_data,EquityIS-method}
\alias{get_data,EquityIS-method}
\title{'data' slot accessor method for S4 objects of class \linkS4class{EquityIS}.}
\usage{
\S4method{get_data}{EquityIS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityIS}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{name}: income statement Bloomberg field names for which data has been found.}
    \item{\code{date}: observation date.}
    \item{\code{value}: corresponding observation.}
  }
}
\description{
'data' slot accessor method for S4 objects of class \linkS4class{EquityIS}.
}
