% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_data,EquityMarket-method}
\alias{get_data,EquityMarket-method}
\title{'data' slot accessor method for S4 objects of class \linkS4class{EquityMarket}.}
\usage{
\S4method{get_data}{EquityMarket}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityMarket}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{field}: equity Bloomberg market data fields for which data has been found.}
    \item{\code{date}: observation date.}
    \item{\code{value}: corresponding value.}
  }
}
\description{
'data' slot accessor method for S4 objects of class \linkS4class{EquityMarket}.
}
