% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{get_data}
\alias{get_data}
\title{Generic method for accessing the 'data' slot of various S4 objects from the \code{pullit} package.}
\usage{
get_data(object)
}
\arguments{
\item{object}{an S4 object with a 'data' slot.}
}
\description{
Generic method for accessing the 'data' slot of various S4 objects from the \code{pullit} package.
}
