% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_fields,BBGHistorical-method}
\alias{get_fields,BBGHistorical-method}
\title{'fields' slot accessor method for S4 objects of class \linkS4class{BBGHistorical}.}
\usage{
\S4method{get_fields}{BBGHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{BBGHistorical}.}
}
\description{
'fields' slot accessor method for S4 objects of class \linkS4class{BBGHistorical}.
}
