% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_fields,EquityCF-method}
\alias{get_fields,EquityCF-method}
\title{'fields' slot accessor method for S4 objects of class \linkS4class{EquityCF}.}
\usage{
\S4method{get_fields}{EquityCF}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityCF}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{section}: cash flow statement sections ('operating', 'financing', 'investing',
      'total') for which data has been found.}
    \item{\code{name}: cash flow statement Bloomberg field names for which data has been found.}
  }
}
\description{
'fields' slot accessor method for S4 objects of class \linkS4class{EquityCF}.
}
