% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_fields,EquityMarket-method}
\alias{get_fields,EquityMarket-method}
\title{'fields' slot accessor method for S4 objects of class \linkS4class{EquityMarket}.}
\usage{
\S4method{get_fields}{EquityMarket}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityMarket}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{field}: equity Bloomberg market data fields for which data has been found.}
  }
}
\description{
'fields' slot accessor method for S4 objects of class \linkS4class{EquityMarket}.
}
