% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_fields,EquityRatios-method}
\alias{get_fields,EquityRatios-method}
\title{'fields' slot accessor method for S4 objects of class \linkS4class{EquityRatios}.}
\usage{
\S4method{get_fields}{EquityRatios}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityRatios}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{type}: financial ratio types ('profitability', 'margin', 'asset turnover',
      'short term liquidity', 'long term solvency') for which data has been found.}
    \item{\code{name}: financial ratio Bloomberg data field names for which data has been found.}
  }
}
\description{
'fields' slot accessor method for S4 objects of class \linkS4class{EquityRatios}.
}
