% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_fields,FuturesCFTC-method}
\alias{get_fields,FuturesCFTC-method}
\title{'fields' slot accessor method for S4 objects of class \linkS4class{FuturesCFTC}.}
\usage{
\S4method{get_fields}{FuturesCFTC}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesCFTC}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{format}: report formats ('legacy', 'disaggregated', 'supplemental' or 'traders in financial futures')
      for which data has been found.}
    \item{\code{underlying}: underlying instruments ('futures only' or 'futures & options') for which data has been found..}
    \item{\code{unit}: counting units (number of 'contracts', 'traders' or 'total') for which data has been found.}
    \item{\code{participant}: trader classifications for which data has been found. Report specific:
      \itemize{
        \item{legacy: 'commercial', 'non-commercial', 'non-reportable', 'total'.}
        \item{disaggregated: 'managed money', 'producer/merchant/processor/user', 'swap dealers', 'other reportables'.}
        \item{supplemental: 'commercial - non-CIT', 'non-commercial - non-CIT', 'index traders - non-CIT', 'index traders'.}
        \item{traders in financial futures: 'asset manager/institutional', 'dealer/intermediary', 'leveraged funds', 'other reportables'.}
      }
    }
    \item{\code{position}: trader positions for which data has been found. Participant specific.
      \itemize{
        \item{commercial: 'long', 'short', 'net'.}
        \item{non-commercial: 'long', 'short', 'net', 'spreading'.}
        \item{non-reportable: 'long', 'short', 'net'.}
        \item{total: 'long', 'short', 'net', 'open interest', 'total'.}
        \item{managed money: 'long', 'short', 'net', 'spreading'.}
        \item{producer/merchant/processor/user: 'long', 'short', 'net'.}
        \item{swap dealers: 'long', 'short', 'net', 'spreading'.}
        \item{other reportables: 'long', 'short', 'net', 'spreading'.}
        \item{commercial - non-CIT: 'long', 'short', 'net'.}
        \item{non-commercial - non-CIT: 'long', 'short', 'net', 'spreading'.}
        \item{index traders - non-CIT: 'long', 'short'.}
        \item{index traders: 'long', 'short', 'net'.}
        \item{asset manager/institutional: 'long', 'short', 'net', 'spreading'.}
        \item{dealer/intermediary: 'long', 'short', 'net', 'spreading'.}
        \item{leveraged funds: 'long', 'short', 'net', 'spreading'.}
        \item{other reportables: 'long', 'short', 'net', 'spreading'.}
      }
    }
  }
}
\description{
'fields' slot accessor method for S4 objects of class \linkS4class{FuturesCFTC}.
}
