% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_fields,FuturesTS-method}
\alias{get_fields,FuturesTS-method}
\title{'fields' slot accessor method for S4 objects of class \linkS4class{FuturesTS}.}
\usage{
\S4method{get_fields}{FuturesTS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesTS}.}
}
\value{
A tibble. Columns include:
   \itemize{
     \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
     \item{\code{TS position}: corresponding futures chain term structure positions for which data has been found.}
     \item{\code{field}: futures Bloomberg market data fields for which data has been found.}
   }
}
\description{
'fields' slot accessor method for S4 objects of class \linkS4class{FuturesTS}.
}
