% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{get_fields}
\alias{get_fields}
\title{Generic method for accessing the 'fields' slot of various S4 objects from the \code{pullit} package.}
\usage{
get_fields(object)
}
\arguments{
\item{object}{an S4 object with a 'fields' slot.}
}
\description{
Generic method for accessing the 'fields' slot of various S4 objects from the \code{pullit} package.
}
