% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_periods,EquityBS-method}
\alias{get_periods,EquityBS-method}
\title{'periods' method for S4 objects of class \linkS4class{EquityBS}.}
\usage{
\S4method{get_periods}{EquityBS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityBS}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{section}: balance sheet sections for which data has been found.}
    \item{\code{subsection}: balance sheet subsections for which data has been found.}
    \item{\code{name}: balance sheet Bloomberg data field names for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }
}
\description{
'periods' method for S4 objects of class \linkS4class{EquityBS}.
}
