% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_periods,EquityCF-method}
\alias{get_periods,EquityCF-method}
\title{'periods' method for S4 objects of class \linkS4class{EquityCF}.}
\usage{
\S4method{get_periods}{EquityCF}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityCF}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{section}: cash flow statement sections for which data has been found.}
    \item{\code{name}: cash flow statement Bloomberg field names for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }
}
\description{
'periods' method for S4 objects of class \linkS4class{EquityCF}.
}
