% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_periods,EquityIS-method}
\alias{get_periods,EquityIS-method}
\title{'periods' method for S4 objects of class \linkS4class{EquityIS}.}
\usage{
\S4method{get_periods}{EquityIS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityIS}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{name}: income statement Bloomberg field names for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }
}
\description{
'periods' method for S4 objects of class \linkS4class{EquityIS}.
}
