% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_periods,FuturesAggregate-method}
\alias{get_periods,FuturesAggregate-method}
\title{'periods' method for S4 objects of class \linkS4class{FuturesAggregate}.}
\usage{
\S4method{get_periods}{FuturesAggregate}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesAggregate}.}
}
\value{
A tibble. Columns include:
   \itemize{
     \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
     \item{\code{field}: futures Bloomberg aggregate data fields for which data has been found.}
     \item{\code{start}: oldest date with observed data.}
     \item{\code{end}: most recent date observed with data.}
   }
}
\description{
'periods' method for S4 objects of class \linkS4class{FuturesAggregate}.
}
