% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_periods,FuturesTS-method}
\alias{get_periods,FuturesTS-method}
\title{'periods' method for S4 objects of class \linkS4class{FuturesTS}.}
\usage{
\S4method{get_periods}{FuturesTS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesTS}.}
}
\value{
A tibble. Columns include:
   \itemize{
     \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
     \item{\code{TS position}: futures term structure positions for which data has been found.}
     \item{\code{field}: futures market data fields for which data has been found.}
     \item{\code{start}: oldest date with observed data.}
     \item{\code{end}: most recent date with observed data.}
   }
}
\description{
'periods' method for S4 objects of class \linkS4class{FuturesTS}.
}
