% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.R
\docType{methods}
\name{get_periods}
\alias{get_periods}
\alias{get_periods,DataHistorical-method}
\alias{get_periods,DataHistorical}
\alias{get_periods,FuturesMarket-method}
\alias{get_periods,DataHistorical,FuturesMarket}
\alias{get_periods,FuturesCFTC-method}
\alias{get_periods,DataHistorical,FuturesCFTC}
\title{Show the time periods for which various S4 objects output of \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}
  contain data.}
\usage{
get_periods(object)

\S4method{get_periods}{DataHistorical}(object)

\S4method{get_periods}{FuturesMarket}(object)

\S4method{get_periods}{FuturesCFTC}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataHistorical} or childs.}
}
\value{
A \link[data.table]{data.table} showing for each Bloomberg ticker and Bloomberg field
  the time periods for which the \code{object} provided contains data.

A data.table showing, for each equity ticker the start and end
  dates of the historical data time series contained in the object.

A data.table showing, for each futures active contract (and term structure contract for
  \linkS4class{FuturesTS} objects) and each field the start and end dates of the historical
  data time series contained in the object.

A data.table showing, for each active contract and position tickers
  the start and end dates of the historical data time series contained in the object.
}
\description{
Generic method for showing the time periods for which various S4 objects from the
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package contain data.
}
