% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{get_periods}
\alias{get_periods}
\title{Generic method for accessing the 'periods' for which various S4 objects from the \code{pullit} package contain data.}
\usage{
get_periods(object)
}
\arguments{
\item{object}{an S4 object with a 'periods' slot.}
}
\description{
Generic method for accessing the 'periods' for which various S4 objects from the \code{pullit} package contain data.
}
