% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_tickers,EquityRatios-method}
\alias{get_tickers,EquityRatios-method}
\title{'tickers' slot accessor method for S4 objects of class \linkS4class{EquityRatios}.}
\usage{
\S4method{get_tickers}{EquityRatios}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityRatios}.}
}
\value{
A character vector: equity Bloomberg tickers for which data has been found.
}
\description{
'tickers' slot accessor method for S4 objects of class \linkS4class{EquityRatios}.
}
