% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.R
\docType{methods}
\name{get_tickers}
\alias{get_tickers}
\alias{get_tickers,BBGHistorical-method}
\alias{get_tickers,BBGHistorical,ANY-method}
\title{Generic method for accessing the 'tickers' slot of various S4 objects from the \code{pullit} package.}
\usage{
get_tickers(object)

\S4method{get_tickers}{BBGHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{BBGHistorical} or childs.}
}
\value{
A tibble.
}
\description{
Generic method for accessing the 'tickers' slot of various S4 objects from the \code{pullit} package.
}
