% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{get_tickers}
\alias{get_tickers}
\title{Generic method for accessing the 'tickers' slot of various S4 objects from the \code{pullit} package.}
\usage{
get_tickers(object)
}
\arguments{
\item{object}{an S4 object with a 'tickers' slot.}
}
\description{
Generic method for accessing the 'tickers' slot of various S4 objects from the \code{pullit} package.
}
