% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{periods,EquityMarket-method}
\alias{periods,EquityMarket-method}
\title{'periods' method for S4 objects of class \linkS4class{EquityMarket}.}
\usage{
\S4method{periods}{EquityMarket}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityMarket}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{field}: equity Bloomberg market data fields for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }
}
\description{
'periods' method for S4 objects of class \linkS4class{EquityMarket}.
}
