% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{periods,EquityRatios-method}
\alias{periods,EquityRatios-method}
\title{'periods' method for S4 objects of class \linkS4class{EquityRatios}.}
\usage{
\S4method{periods}{EquityRatios}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityRatios}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{type}: financial ratio types for which data has been found.}
    \item{\code{name}: financial ratio Bloomberg data field names for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }
}
\description{
'periods' method for S4 objects of class \linkS4class{EquityRatios}.
}
