% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{periods,FuturesCFTC-method}
\alias{periods,FuturesCFTC-method}
\title{'periods' method for S4 objects of class \linkS4class{FuturesCFTC}.}
\usage{
\S4method{periods}{FuturesCFTC}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesCFTC}.}
}
\value{
A tibble. Columns include:
  \itemize{
    \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
    \item{\code{format}: report formats for which data has been found.}
    \item{\code{underlying}: underlying instruments for which data has been found..}
    \item{\code{unit}: counting units for which data has been found.}
    \item{\code{participant}: trader classifications for which data has been found.}
    \item{\code{position}: trader positions for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date observed with data.}
  }
}
\description{
'periods' method for S4 objects of class \linkS4class{FuturesCFTC}.
}
