% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\name{periods}
\alias{periods}
\title{Generic method for accessing the 'periods' for which various S4 objects from the \code{pullit} package contain data.}
\usage{
periods(object)
}
\arguments{
\item{object}{an S4 object with a 'periods' slot.}
}
\description{
Generic method for accessing the 'periods' for which various S4 objects from the \code{pullit} package contain data.
}
