% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot_TS,FuturesTS-method}
\alias{plot_TS,FuturesTS-method}
\alias{plot_TS,BBGHistorical,FuturesTS}
\title{Plot futures term structure}
\usage{
\S4method{plot_TS}{FuturesTS}(object, ticker)
}
\description{
Plot futures term structure
}
