% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{plot_TS}
\alias{plot_TS}
\title{Generic method for plotting historical futures term structure data contained in an S4 object
  of class \linkS4class{FuturesTS} from the \code{pullit} package.}
\usage{
plot_TS(object, ticker)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesTS}.}

\item{ticker}{a scalar character vector. Active contract Bloomberg ticker to plot the
term structure for.}
}
\description{
Generic method for plotting historical futures term structure data contained in an S4 object
  of class \linkS4class{FuturesTS} from the \code{pullit} package.
}
