% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{pull_equity_book}
\alias{pull_equity_book}
\title{Retrieves equity historical book data.}
\usage{
pull_equity_book(source = "Bloomberg", book, tickers, start, end,
  verbose = T, file = NULL, ...)
}
\arguments{
\item{source}{a scalar character vector. Specifies the data source for the query:
"Bloomberg" or "storethat". Defaults to "Bloomberg".}

\item{book}{a scalar chatacter vector, 'key stats', 'income statement', 'balance sheet',
'cash flow statement' or 'ratios'.}

\item{tickers}{a chatacter vector. Specifies the Bloomberg equity tickers to query data for.}

\item{start}{a scalar character vector. Specifies the starting date for the query in the
following format: 'yyyy-mm-dd'.}

\item{end}{a scalar character vector. Specifies the end date for the query in the
following format: 'yyyy-mm-dd'.}

\item{verbose}{a logical scalar vector. Should progression messages be printed?
Defaults to TRUE.}

\item{file}{a scalar chatacter vector. Optional parameter that specifies the
target \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite
database file to retrieve data from.}

\item{...}{optional parameters to pass to the \link[Rblpapi]{bdh} function from the
\href{http://dirk.eddelbuettel.com/code/rblpapi.html}{\pkg{Rblpapi}} package used
  for the query (\code{options} parameter).}
}
\value{
An S4 object of class \linkS4class{EquityKS} (\code{book = 'key stats'}),
  \linkS4class{EquityIS} (\code{book = 'income statement'}), \\linkS4class{EquityBS}
  (\code{book = 'balance sheet'}), \linkS4class{EquityCF}
  (\code{book = 'cash flow statement'}), \linkS4class{EquityRatios} (\code{book = 'ratios'}).
}
\description{
Retrieves equity historical book data from Bloomberg. Books include
  'key stats', 'income statement', 'balance sheet', 'cash flow statement' and
  'ratios' or retrieves it from an existing
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.
}
\examples{
\dontrun{

    BBG_KS <- pull_equity_books(source = "Bloomberg", book = 'key stats',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"), start = "2000-01-01",
      end = as.character(Sys.Date()))

    BBG_PL <- pull_equity_books(source = "Bloomberg", book = 'income statement',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"), start = "2000-01-01",
      end = as.character(Sys.Date()))

    BBG_BS <- pull_equity_books(source = "Bloomberg", book = 'balance sheet',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"), start = "2000-01-01",
      end = as.character(Sys.Date()))

    BBG_CF <- pull_equity_books(source = "Bloomberg", book = 'cash flow statement',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"), start = "2000-01-01",
      end = as.character(Sys.Date()))

    BBG_R <- pull_equity_books(source = "Bloomberg", book = 'ratios',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"), start = "2000-01-01",
      end = as.character(Sys.Date()))

    storethat_KS <- pull_equity_books(source = "storethat", file = "~/storethat.sqlite",
      book = 'key stats', tickers = c("BP/ LN Equity", "WEIR LN Equity"),
      start = "2000-01-01", end = as.character(Sys.Date()))

    storethat_PL <- pull_equity_books(source = "storethat", book = 'income statement',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"), start = "2000-01-01",
      end = as.character(Sys.Date()))

    storethat_BS <- pull_equity_books(source = "storethat", file = "~/storethat.sqlite",
      book = 'balance sheet', tickers = c("BP/ LN Equity", "WEIR LN Equity"),
      start = "2000-01-01", end = as.character(Sys.Date()))

    storethat_CF <- pull_equity_books(source = "storethat", book = 'cash flow statement',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"), start = "2000-01-01",
      end = as.character(Sys.Date()))

    storethat_R <- pull_equity_books(source = "storethat",  file = "~/storethat.sqlite",
      book = 'ratios', tickers = c("BP/ LN Equity", "WEIR LN Equity"),
      start = "2000-01-01", end = as.character(Sys.Date()))


  }


}
\seealso{
\itemize{

    \item{The \link[BBGsymbols]{fields} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
     (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite) for
     details on the Bloomnerg fields used here.}

    \item{Helper datasets in the
    \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} and
    \href{https://github.com/bautheac/GICS/}{\pkg{GICS}} packages
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}

  }
}
