% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{pull_futures_CFTC}
\alias{pull_futures_CFTC}
\title{Retrieves futures CFTC reports historical data.}
\usage{
pull_futures_CFTC(
  source = "Bloomberg",
  active_contract_tickers,
  start,
  end,
  verbose = T,
  file = NULL,
  ...
)
}
\arguments{
\item{source}{a scalar character vector. Specifies the data source for the query:
"Bloomberg" or "storethat". Defaults to "Bloomberg".}

\item{active_contract_tickers}{a chatacter vector. Specifies the futures active
contract Bloomberg tickers to query data for.}

\item{start}{a scalar character vector. Specifies the starting date for the query
in the following format: 'yyyy-mm-dd'.}

\item{end}{a scalar character vector. Specifies the end date for the query in the
following format: 'yyyy-mm-dd'.}

\item{verbose}{a logical scalar vector. Should progression messages be printed?
Defaults to TRUE.}

\item{file}{a scalar chatacter vector. Optional parameter that specifies the
target \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite
database file to retrieve data from.}

\item{...}{optional parameters to pass to the \link[Rblpapi]{bdh} function from the
\href{http://dirk.eddelbuettel.com/code/rblpapi.html}{\pkg{Rblpapi}} package used
  for the query (\code{options} parameter).}
}
\value{
An S4 object of class \linkS4class{FuturesCFTC}.
}
\description{
Provided with a set of futures active contract Bloomberg tickers and a
  time period, queries Bloomberg for the corresponding futures CFTC reports
  historical data or retrieves it from an
  existing \href{https://github.com/bautheac/storethat/}{\pkg{storethat}}
  SQLite database.
}
\examples{
\dontrun{

    BBG_CFTC <- pull_futures_CFTC(source = "Bloomberg",
      active_contract_tickers = c("W A Comdty", "KWA Comdty"),
      start = "2010-01-01", end = as.character(Sys.Date()))

    storethat_CFTC <- pull_futures_CFTC(source = "storethat",
      active_contract_tickers = c("W A Comdty", "KWA Comdty"),
      start = "2010-01-01", end = as.character(Sys.Date()))

  }


}
\seealso{
\itemize{

    \item{"GFUT <GO>" on a Bloomberg terminal.}

    \item{The \link[BBGsymbols]{tickers_cftc} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite)
    for details on the Bloomnerg position tickers used here.}

    \item{Helper datasets in the
    \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} package
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}

  }
}
