% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{pull_index_info}
\alias{pull_index_info}
\title{Retrieves index qualitative information.}
\usage{
pull_index_info(source = "Bloomberg", tickers, file = NULL, ...)
}
\arguments{
\item{source}{a scalar character vector. Specifies the data source for the query:
"Bloomberg" or "storethat". Defaults to "Bloomberg".}

\item{tickers}{a chatacter vector. Specifies the futures active contract Bloomberg
tickers to query data for.}

\item{file}{a scalar chatacter vector. Optional parameter that specifies the
target \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite
database file to retrieve data from.}

\item{...}{optional parameters to pass to the \link[Rblpapi]{bdp} function from the
\href{http://dirk.eddelbuettel.com/code/rblpapi.html}{\pkg{Rblpapi}} package used
  for the query (\code{options} parameter).}
}
\value{
An S4 object of class \linkS4class{IndexInfo}.
}
\description{
Provided with a set of index Bloomberg tickers pulls qualitative
  information on the corresponding index(es) or
  retrieves it from an existing
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.
}
\examples{
\dontrun{

    BBG_info <- pull_index_info(source = "Bloomberg", tickers = c("NEIXCTA Index", "BARCCTA Index"))

    storethat_info <- pull_index_info(source = "storethat", file = "~/storethat.sqlite",
      tickers = c("NEIXCTA Index", "BARCCTA Index"))

    storethat_info <- pull_index_info(source = "storethat", tickers = c("NEIXCTA Index", "BARCCTA Index"))

  }


}
\seealso{
\itemize{

    \item{The \link[BBGsymbols]{fields} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
     (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite) for details
     on the Bloomnerg fields used here.}

    \item{Helper datasets in the
    \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} and
    \href{https://github.com/bautheac/GICS/}{\pkg{GICS}} packages
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}


  }
}
