% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{random_futures_spot}
\alias{random_futures_spot}
\title{Creates a FuturesSpot object from futures spot data.}
\usage{
random_futures_spot(data)
}
\arguments{
\item{data}{a dataframe containing futures spot data. Columns should include:

  \itemize{
    \item{ticker: futures active contract Bloomberg tickers.}
    \item{field: field Bloomberg symbols. See \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} for a list of accepted
    symbols.}
    \item{date: observation date (YYYY-MM-DD).}
    \item{value: observation value.}
  }}
}
\value{
An S4 object of class \linkS4class{FuturesSpot}.
}
\description{
Futures spot data refers to data on the corresponding cash market. Futures spot/cash data
  is not avaiable on Bloomberg at the time of writing. Should spot data be retrieved from other data
  source(s), this function would creates a FuturesSpot data object from it in order to facilitates
  storage in a \href{https://bautheac.github.io/storethat/}{\pkg{storethat}} database.
}
