% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{random_index_market}
\alias{random_index_market}
\title{Creates an IndexMarket object from index market data contained in a dataframe.}
\usage{
random_index_market(data)
}
\arguments{
\item{data}{a dataframe containing index market data. Columns should include:

  \itemize{
    \item{ticker: index Bloomberg tickers.}
    \item{field: field Bloomberg symbols.
      See \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} for a list
      of accepted symbols.}
    \item{date: observation date (YYYY-MM-DD).}
    \item{value: observation value.}
  }}
}
\value{
An S4 object of class \linkS4class{IndexMarket}.
}
\description{
Should index data be retrieved from other source(s) than Bloomberg,
  this function creates a IndexMarket data object from it in order to facilitates
  storage in a \href{https://bautheac.github.io/storethat/}{\pkg{storethat}} database.
}
