% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,BBGHistorical-method}
\alias{show,BBGHistorical-method}
\title{Show method for S4 object of class \code{BBGHistorical}.}
\usage{
\S4method{show}{BBGHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{BBGHistorical}.}
}
\description{
Show method for S4 object of class \code{BBGHistorical}.
}
