% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,DataInfo-method}
\alias{show,DataInfo-method}
\title{Show method for S4 object of class \linkS4class{DataInfo}.}
\usage{
\S4method{show}{DataInfo}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataInfo}.}
}
\description{
Show method for S4 object of class \linkS4class{DataInfo}.
}
