% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{show,EquityBS-method}
\alias{show,EquityBS-method}
\title{Show method for S4 object of class \code{EquityBS}.}
\usage{
\S4method{show}{EquityBS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityBS}.}
}
\description{
Show method for S4 object of class \code{EquityBS}.
}
