% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{show,EquityCF-method}
\alias{show,EquityCF-method}
\title{Show method for S4 object of class \code{EquityCF}.}
\usage{
\S4method{show}{EquityCF}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityCF}.}
}
\description{
Show method for S4 object of class \code{EquityCF}.
}
