% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,EquityHistorical-method}
\alias{show,EquityHistorical-method}
\title{Show method for S4 object of class \linkS4class{EquityHistorical}.}
\usage{
\S4method{show}{EquityHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityHistorical}.}
}
\description{
Show method for S4 object of class \linkS4class{EquityHistorical}.
}
