% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{show,EquityIS-method}
\alias{show,EquityIS-method}
\title{Show method for S4 object of class \code{EquityIS}.}
\usage{
\S4method{show}{EquityIS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityIS}.}
}
\description{
Show method for S4 object of class \code{EquityIS}.
}
