% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{show,EquityMarket-method}
\alias{show,EquityMarket-method}
\title{Show method for S4 object of class \code{EquityMarket}.}
\usage{
\S4method{show}{EquityMarket}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityMarket}.}
}
\description{
Show method for S4 object of class \code{EquityMarket}.
}
