% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,EquityRatios-method}
\alias{show,EquityRatios-method}
\title{Show method for S4 object of class \code{EquityRatios}.}
\usage{
\S4method{show}{EquityRatios}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityRatios}.}
}
\description{
Show method for S4 object of class \code{EquityRatios}.
}
