% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,FundHistorical-method}
\alias{show,FundHistorical-method}
\title{Show method for S4 object of class \linkS4class{FundHistorical}.}
\usage{
\S4method{show}{FundHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FundHistorical}.}
}
\description{
Show method for S4 object of class \linkS4class{FundHistorical}.
}
