% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,FuturesAggregate-method}
\alias{show,FuturesAggregate-method}
\title{Show method for S4 object of class \code{FuturesAggregate}.}
\usage{
\S4method{show}{FuturesAggregate}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesAggregate}.}
}
\description{
Show method for S4 object of class \code{FuturesAggregate}.
}
