% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,FuturesCFTC-method}
\alias{show,FuturesCFTC-method}
\title{Show method for S4 object of class \linkS4class{FuturesCFTC}..}
\usage{
\S4method{show}{FuturesCFTC}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesCFTC}.}
}
\description{
Show method for S4 object of class \linkS4class{FuturesCFTC}..
}
