% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,FuturesHistorical-method}
\alias{show,FuturesHistorical-method}
\title{Show method for S4 object of class \linkS4class{FuturesHistorical}.}
\usage{
\S4method{show}{FuturesHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesHistorical}.}
}
\description{
Show method for S4 object of class \linkS4class{FuturesHistorical}.
}
