% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,FuturesTS-method}
\alias{show,FuturesTS-method}
\title{Show method for S4 object of class \code{FuturesTS}.}
\usage{
\S4method{show}{FuturesTS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesTS}.}
}
\description{
Show method for S4 object of class \code{FuturesTS}.
}
