% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,DataInfo-method}
\alias{show,DataInfo-method}
\alias{show,DataInfo}
\alias{show,FuturesHistorical-method}
\alias{show,FuturesHistorical}
\alias{show,EquityHistorical-method}
\alias{show,EquityHistorical}
\alias{show,FundHistorical-method}
\alias{show,FundHistorical}
\alias{show,FuturesTS-method}
\alias{show,FuturesTS}
\alias{show,FuturesCFTC-method}
\alias{show,FuturesCFTC}
\title{Show method for S4 object of class \linkS4class{DataInfo}.}
\usage{
\S4method{show}{DataInfo}(object)

\S4method{show}{FuturesHistorical}(object)

\S4method{show}{EquityHistorical}(object)

\S4method{show}{FundHistorical}(object)

\S4method{show}{FuturesTS}(object)

\S4method{show}{FuturesCFTC}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataInfo}.}

\item{object}{an S4 object of class \linkS4class{FuturesHistorical}.}

\item{object}{an S4 object of class \linkS4class{EquityHistorical}.}

\item{object}{an S4 object of class \linkS4class{FundHistorical}.}

\item{object}{an S4 object of class \linkS4class{FuturesTS}.}

\item{object}{an S4 object of class \linkS4class{FuturesCFTC}.}
}
\description{
Show method for S4 object of class \linkS4class{DataInfo}.

Show method for S4 object of class \linkS4class{FuturesHistorical}.

Show method for S4 object of class \linkS4class{EquityHistorical}.

Show method for S4 object of class \linkS4class{FundHistorical}.

Show method for S4 object of class \linkS4class{FuturesTS}..

Show method for S4 object of class \linkS4class{FuturesCFTC}..
}
