% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{storethat_equity_book}
\alias{storethat_equity_book}
\title{Retrieves equity historical book data from a
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.}
\usage{
storethat_equity_book(file = NULL, book, tickers, start, end, verbose = T)
}
\arguments{
\item{file}{a scalar chatacter vector. Specifies the target \pkg{storethat}
SQLite database file.}

\item{book}{a scalar chatacter vector, 'key stats', 'income statement',
'balance sheet', 'cash flow statement' or 'ratios'.}

\item{tickers}{a chatacter vector. Specifies the Bloomberg equity tickers
to query data for.}

\item{start}{a scalar character vector. Specifies the starting date for the query
in the following format: 'yyyy-mm-dd'.}

\item{end}{a scalar character vector. Specifies the end date for the query in the
following format: 'yyyy-mm-dd'.}

\item{verbose}{a logical scalar vector. Should progression messages be printed?
Defaults to TRUE.}
}
\value{
An S4 object of class \linkS4class{EquityKS} (\code{book = 'key stats'}),
  \linkS4class{EquityIS} (\code{book = 'income statement'}), \\linkS4class{EquityBS}
  (\code{book = 'balance sheet'}), \linkS4class{EquityCF}
  (\code{book = 'cash flow statement'}), \linkS4class{EquityRatios}
  (\code{book = 'ratios'}).
}
\description{
Retrieve equity historical book data previously stored in a
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.
  Books include 'key stats', 'income statement', 'balance sheet',
  'cash flow statement' and 'ratios'.
}
\examples{
\dontrun{

    storethat_KS <- storethat_equity_books(book = 'key stats',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"),
      start = "2010-01-01", end = as.character(Sys.Date()))

    storethat_PL <- storethat_equity_books(book = 'income statement',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"),
      start = "2010-01-01", end = as.character(Sys.Date()))

    storethat_BS <- storethat_equity_books(book = 'balance sheet',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"),
      start = "2010-01-01", end = as.character(Sys.Date()))

    storethat_CF <- storethat_equity_books(book = 'cash flow statement',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"),
      start = "2010-01-01", end = as.character(Sys.Date()))

    storethat_R <- storethat_equity_books(book = 'ratios',
      tickers = c("BP/ LN Equity", "WEIR LN Equity"),
      start = "2010-01-01", end = as.character(Sys.Date()))

  }

}
\seealso{
\itemize{

    \item{The \link[BBGsymbols]{fields} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite) for details
    on the Bloomnerg fields used here.}

    \item{Helper datasets in the
    \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} and
    \href{https://github.com/bautheac/GICS/}{\pkg{GICS}} packages
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}

    \item{The \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite,
    in particular the \href{https://github.com/bautheac/storethat/}{\pkg{storethat}}
    package.}

  }
}
