% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{storethat_equity_books}
\alias{storethat_equity_books}
\title{Equity historical book data from Bloomberg}
\usage{
storethat_equity_books(book = "key stats", tickers = "NEM US Equity",
  start = "2018-01-01", end = "2018-06-30", verbose = TRUE,
  file = NULL)
}
\arguments{
\item{book}{A scalar chatacter vector, 'key stats', 'income statement', 'balance sheet', 'cash flow statement' or 'ratios'.
Defaults to 'key stats' which returns historical data for a set of key financial indicators; a quick dynamic financial
snapshot of the selected names.}

\item{tickers}{A chatacter vector. Specifies the Bloomberg equity tickers to query data for.
Defaults to 'NEM US Equity', the Bloomberg equity ticker for the Newmont Minning Corporation.}

\item{start}{A scalar character vector. Specifies the starting date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-01-01'.}

\item{end}{A scalar character vector. Specifies the end date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-06-30'.}

\item{verbose}{A logical scalar vector. Should progression messages be printed? Defaults to TRUE.}

\item{file}{A scalar chatacter vector. Specifies the target \code{storethat} SQLite database file.}
}
\value{
An S4 object of class \code{\linkS4class{EquityKS}} (\code{book = 'key stats'}), \code{\linkS4class{EquityIS}}
  (\code{book = 'income statement'}), \code{\linkS4class{EquityBS}} (\code{book = 'balance sheet'}),
   \code{\linkS4class{EquityCF}} (\code{book = 'cash flow statement'}), \code{\linkS4class{EquityRatios}}
   (\code{book = 'ratios'}). Slots include:
  \itemize{
    \item{\code{tickers}: a data.table. Equity Bloomberg tickers for which data has been found.}
    \item{\code{fields}: a data.table. Data fields for which data has been found.}
    \item{\code{data}: a data.table The data retrieved in the query.}
    \item{\code{call}: a scalar character vector. The original call to the constructor.}
  }
}
\description{
Retrieve equity historical book data from Bloomberg. Books include 'key stats', 'income statement', 'balance sheet',
  'cash flow statement' and 'ratios' with the corresponding data fields retrieved from the 'fields' dataset of the \code{bbgsymbols}
  package.
}
\examples{
\dontrun{
  bbg_equity_books(book = 'key stats')
  bbg_equity_books(book = 'income statement')
  bbg_equity_books(book = 'balance sheet')
  bbg_equity_books(book = 'cash flow statement')
  bbg_equity_books(book = 'ratios')
}

}
\seealso{
The \code{\link[bbgsymbols]{fields}} dataset in the \code{bbgsymbols} package for details on the Bloomnerg fields used here.
}
