% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{storethat_fund_info}
\alias{storethat_fund_info}
\title{Fund qualitative information from a \href{https://github.com/bautheac/storethat/}{\pkg{storethat}}
  SQLite database}
\usage{
storethat_fund_info(tickers, file = NULL)
}
\arguments{
\item{tickers}{a chatacter vector. Specifies the futures active contract Bloomberg tickers to
query data for.}

\item{file}{a scalar chatacter vector. Specifies the target \pkg{storethat} SQLite database file.}
}
\value{
An S4 object of class \linkS4class{FundInfo}.
}
\description{
Provided with a set of equity Bloomberg tickers retrieves qualitative information
  on the corresponding fund previously stored in a
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.
}
\examples{
\dontrun{

    storethat_fund_info(tickers = c("SPY US Equity", "IVV US Equity"))

  }


}
\seealso{
\itemize{

    \item{The \link[BBGsymbols]{fields} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
     (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite) for details on the Bloomnerg
     fields used here.}

    \item{Helper datasets in the \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} and
    \href{https://github.com/bautheac/GICS/}{\pkg{GICS}} packages
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}

    \item{The \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite,
    in particular the \href{https://github.com/bautheac/storethat/}{\pkg{storethat}}
    package.}

  }
}
