% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{storethat_fund_market}
\alias{storethat_fund_market}
\title{Fund historical market data from a \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database}
\usage{
storethat_fund_market(file = NULL, tickers, start, end, verbose = TRUE)
}
\arguments{
\item{file}{A scalar chatacter vector. Specifies the target \pkg{storethat} SQLite database file.}

\item{tickers}{A chatacter vector. Specifies the Bloomberg fund tickers to query data for.}

\item{start}{A scalar character vector. Specifies the starting date for the query in the following format: 'yyyy-mm-dd'.}

\item{end}{A scalar character vector. Specifies the end date for the query in the following format: 'yyyy-mm-dd'.}

\item{verbose}{A logical scalar vector. Should progression messages be printed? Defaults to TRUE.}
}
\value{
An S4 object of class \linkS4class{FundMarket}.
}
\description{
Provided with a set of Bloomberg fund tickers and a time period, retrieves the corresponding fund
  historical market data previously stored in a \href{https://github.com/bautheac/storethat/}{\pkg{storethat}}
  SQLite database.
}
\examples{
\dontrun{
    storethat_fund_market(tickers = c("SPY US Equity", "IVV US Equity"),
      start = "2000-01-01", end = as.character(Sys.Date()))
  }

}
\seealso{
\itemize{
    \item{The \link[BBGsymbols]{fields} dataset in the \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
     (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite) for details on the Bloomnerg fields used here.}
    \item{Helper datasets in the \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} and
    \href{https://github.com/bautheac/GICS/}{\pkg{GICS}} packages
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}
  }
}
