% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{storethat_futures_info}
\alias{storethat_futures_info}
\title{Futures series qualitative information from a \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database}
\usage{
storethat_futures_info(file = NULL, active_contract_tickers)
}
\arguments{
\item{file}{A scalar chatacter vector. Specifies the target \pkg{storethat} SQLite database file.}

\item{active_contract_tickers}{A chatacter vector. Specifies the futures active contract Bloomberg tickers to query data for.}
}
\value{
An S4 object of class \linkS4class{FuturesInfo}.
}
\description{
Provided with a set of futures active contract Bloomberg tickers retrieves qualitative information
  on the corresponding futures series previously stored in a \href{https://github.com/bautheac/storethat/}{\pkg{storethat}}
  SQLite database.
}
\examples{
\dontrun{
    storethat_futures_info(active_contract_tickers = c("W A Comdty", "KWA Comdty"))
  }

}
\seealso{
\itemize{
    \item{The \link[BBGsymbols]{fields} (\pkg{finRes} suite) dataset in the \pkg{BBGsymbols} package for details on the Bloomnerg fields used here.}
    \item{Helper datasets in the \pkg{fewISOs} and \pkg{GICS} packages (\pkg{finRes} suite).}
  }
}
