% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{storethat_futures_spot}
\alias{storethat_futures_spot}
\title{Retrieves futures spot historical data from from a
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.}
\usage{
storethat_futures_spot(file, active_contract_tickers, start, end, verbose)
}
\arguments{
\item{file}{a scalar chatacter vector. Specifies the target
\href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database file.}

\item{active_contract_tickers}{a chatacter vector. Specifies the futures active contract
Bloomberg tickers to query data for.}

\item{start}{a scalar character vector. Specifies the starting date for the query in the
following format: 'yyyy-mm-dd'.}

\item{end}{a scalar character vector. Specifies the end date for the query in the
following format: 'yyyy-mm-dd'.}

\item{verbose}{a logical scalar vector. Should progression messages be printed?
Defaults to TRUE.}
}
\value{
An S4 object of class \linkS4class{FuturesSpot}.
}
\description{
Provided with a set of Bloomberg futures active contract tickers and
  a time period, retrieves the corresponding futures historical spot data previously
  stored in a \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite
  database. Futures spot data refers to the corresponding cash market.
}
\examples{
\dontrun{

    storethat_futures_market(type = "spot",
      active_contract_tickers = c("W A Comdty", "KWA Comdty"),
      start = "2000-01-01", end = as.character(Sys.Date()))

  }


}
