% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{storethat_index_market}
\alias{storethat_index_market}
\title{Retrieves index historical market data from a
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.}
\usage{
storethat_index_market(file = NULL, tickers, start, end, verbose = T)
}
\arguments{
\item{file}{a scalar chatacter vector. Specifies the target \pkg{storethat}
SQLite database file.}

\item{tickers}{a chatacter vector. Specifies the Bloomberg index tickers to query
data for.}

\item{start}{a scalar character vector. Specifies the starting date for the query
in the following format: 'yyyy-mm-dd'.}

\item{end}{a scalar character vector. Specifies the end date for the query in the
following format: 'yyyy-mm-dd'.}

\item{verbose}{a logical scalar vector. Should progression messages be printed?
Defaults to TRUE.}
}
\value{
An S4 object of class \linkS4class{indexMarket}.
}
\description{
Provided with a set of Bloomberg index tickers and a time period,
  retrieves the corresponding index historical market data previously stored in a
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.
}
\examples{
\dontrun{

    storethat_index_market(tickers = c("NEIXCTA Index", "BARCCTA Index"),
      start = "2000-01-01", end = as.character(Sys.Date()))

  }


}
\seealso{
\itemize{

    \item{The \link[BBGsymbols]{fields} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
     (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite) for details on the
     Bloomnerg fields used here.}

    \item{Helper datasets in the
    \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} and
    \href{https://github.com/bautheac/GICS/}{\pkg{GICS}} packages
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}

    \item{The \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite,
    in particular the \href{https://github.com/bautheac/storethat/}{\pkg{storethat}}
    package.}

  }
}
