% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{storethat_update}
\alias{storethat_update}
\title{Update your \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite
  database.}
\usage{
storethat_update(instrument = "all", book = "all", name = "all",
  file = NULL, verbose = TRUE)
}
\arguments{
\item{instrument}{a scalar character vector. Specifies the financial instruments to get a
snapshot for. Must be one of 'all', equity', 'fund' or 'futures'.}

\item{book}{a scalar character vector. Instrument dependent; for a given instrument, specifies
the book for the snapshot; 'all' snapshots all the books available for the given instrument.}

\item{name}{a scalar character vector. Instrument dependent; for a given instrument, specifies
the name for the snapshot; 'all' snapshots all the names available for the given instrument.}

\item{file}{a scalar character vector. Specifies the path to the appropriate 'storethat.sqlite'
file.}

\item{verbose}{a logical scalar vector. Should progression messages be printed?
Defaults to TRUE.}
}
\description{
Updates an existing \href{https://github.com/bautheac/storethat/}{\pkg{storethat}}
  SQLite database with fresh data up to date.
}
\details{
For a given financial instrument, updates the database by book and name.
}
\examples{
\dontrun{storethat_update()}


}
\seealso{
The \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite,
  in particular the \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} &
  \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} packages.
}
