% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{tickers,EquityIS-method}
\alias{tickers,EquityIS-method}
\title{'tickers' slot accessor method for S4 objects of class \linkS4class{EquityIS}.}
\usage{
\S4method{tickers}{EquityIS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityIS}.}
}
\value{
A character vector: equity Bloomberg tickers for which data has been found.
}
\description{
'tickers' slot accessor method for S4 objects of class \linkS4class{EquityIS}.
}
