% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{tickers,EquityMarket-method}
\alias{tickers,EquityMarket-method}
\title{'tickers' slot accessor method for S4 objects of class \linkS4class{EquityMarket}.}
\usage{
\S4method{tickers}{EquityMarket}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityMarket}.}
}
\value{
A character vector: equity Bloomberg tickers for which data has been found.
}
\description{
'tickers' slot accessor method for S4 objects of class \linkS4class{EquityMarket}.
}
