% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{tickers,FuturesAggregate-method}
\alias{tickers,FuturesAggregate-method}
\title{'tickers' slot accessor method for S4 objects of class \linkS4class{FuturesAggregate}.}
\usage{
\S4method{tickers}{FuturesAggregate}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesAggregate}.}
}
\value{
A character vector: active contract Bloomberg tickers for which data has been found.
}
\description{
'tickers' slot accessor method for S4 objects of class \linkS4class{FuturesAggregate}.
}
