% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{tickers,FuturesCFTC-method}
\alias{tickers,FuturesCFTC-method}
\title{'tickers' slot accessor method for S4 objects of class \linkS4class{FuturesCFTC}.}
\usage{
\S4method{tickers}{FuturesCFTC}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesCFTC}.}
}
\value{
A character vector: active contract Bloomberg tickers for which data has been found.
}
\description{
'tickers' slot accessor method for S4 objects of class \linkS4class{FuturesCFTC}.
}
