% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{tickers,FuturesTS-method}
\alias{tickers,FuturesTS-method}
\title{'tickers' slot accessor method for S4 objects of class \linkS4class{FuturesTS}.}
\usage{
\S4method{tickers}{FuturesTS}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FuturesTS}.}
}
\value{
A tibble. Columns include:
   \itemize{
     \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
     \item{\code{ticker}: futures term structure Bloomberg tickers for which data has been found.}
     \item{\code{TS position}: corresponding futures term structure position.}
     \item{\code{roll type}: corresponding roll type (by name).}
     \item{\code{roll days}: corresponding day(s) offset.}
     \item{\code{roll months}: corresponding month(s) offset.}
     \item{\code{roll adjustment}: corresponding roll adjustment (by name).}
   }
}
\description{
'tickers' slot accessor method for S4 objects of class \linkS4class{FuturesTS}.
}
