% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{BBG_futures_aggregate}
\alias{BBG_futures_aggregate}
\title{Retrieves futures aggregate historical data from Bloomberg.}
\usage{
BBG_futures_aggregate(active_contract_tickers, start, end, verbose, ...)
}
\arguments{
\item{active_contract_tickers}{a chatacter vector. Specifies the futures active contract
Bloomberg tickers to query data for.}

\item{start}{a scalar character vector. Specifies the starting date for the query in the
following format: 'yyyy-mm-dd'.}

\item{end}{a scalar character vector. Specifies the end date for the query in the following
format: 'yyyy-mm-dd'.}

\item{verbose}{a logical scalar vector. Should progression messages be printed? Defaults to TRUE.}

\item{...}{optional parameters to pass to the \link[Rblpapi]{bdh} function from the
\href{http://dirk.eddelbuettel.com/code/rblpapi.html}{\pkg{Rblpapi}} package used
  for the query (\code{options} parameter).}
}
\value{
An S4 object of class \linkS4class{FuturesAggregate}.
}
\description{
Provided with a set of futures active contract Bloomberg tickers and a time period,
  queries Bloomberg for the corresponding futures historical aggregate data. For each individual field,
  futures aggregate data represents the aggregation of the corresponding field values over all the
  corresponding term structure contracts.
}
\examples{
\dontrun{
    BBG_futures_market(type = "aggregate",
      active_contract_tickers = c("W A Comdty", "KWA Comdty"),
      start = "2000-01-01", end = as.character(Sys.Date()))
  }


}
\seealso{
\itemize{

    \item{"GFUT <GO>" on a Bloomberg terminal.}

    \item{The \link[BBGsymbols]{rolls} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite) for details
    regarding the \code{roll_type} & \code{roll_adjustment} parameters.}
    \item{The \link[BBGsymbols]{fields} dataset in the
    \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} package
    for details on the Bloomnerg fields used here.}

    \item{Helper datasets in the
    \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}} package
    (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite).}

  }
}
