% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.r
\name{bbg_futures_CFTC}
\alias{bbg_futures_CFTC}
\title{Futures CFTC reports historical data from Bloomberg}
\usage{
bbg_futures_CFTC(active_contract_tickers = "C A Comdty",
  start = "2018-01-01", end = "2018-06-30", ...)
}
\arguments{
\item{active_contract_tickers}{A chatacter vector. Specifies the Bloomberg futures active contract tickers to query data for.
Defaults to 'C A Comdty', the Bloomberg active contract ticker for the XCBT corn futures series.}

\item{start}{A scalar character vector. Specifies the starting date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-01-01'.}

\item{end}{A scalar character vector. Specifies the end date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-06-30'.}

\item{...}{Optional parameters to pass to the \code{\link[Rblpapi]{bdh}} function from the \code{Rblpapi} package used
for the query (\code{options} parameter).}
}
\value{
An S4 object of class \code{\linkS4class{FuturesCFTC}} with slots:
  \itemize{
    \item{\code{tickers}: a character vector. Active contract Bloomberg tickers for which data has been found.}
    \item{\code{fields}: a tibble. Columns include:
      \itemize{
        \item{\code{format}: report formats ('legacy', 'disaggregated', 'supplemental' or 'traders in financial futures')
          for which data has been found.}
        \item{\code{underlying}: underlying instruments ('futures only' or 'futures & options') for which data has been found..}
        \item{\code{unit}: counting units (number of 'contracts', 'traders' or 'total') for which data has been found.}
        \item{\code{participant}: trader classifications for which data has been found. Report specific:
          \itemize{
            \item{legacy: 'commercial', 'non-commercial', 'non-reportable', 'total'.}
            \item{disaggregated: 'managed money', 'producer/merchant/processor/user', 'swap dealers', 'other reportables'.}
            \item{supplemental: 'commercial - non-CIT', 'non-commercial - non-CIT', 'index traders - non-CIT', 'index traders'.}
            \item{traders in financial futures: 'asset manager/institutional', 'dealer/intermediary', 'leveraged funds', 'other reportables'.}
          }
        }
        \item{\code{position}: trader positions for which data has been found. Participant specific.
          \itemize{
            \item{commercial: 'long', 'short', 'net'.}
            \item{non-commercial: 'long', 'short', 'net', 'spreading'.}
            \item{non-reportable: 'long', 'short', 'net'.}
            \item{total: 'long', 'short', 'net', 'open interest', 'total'.}
            \item{managed money: 'long', 'short', 'net', 'spreading'.}
            \item{producer/merchant/processor/user: 'long', 'short', 'net'.}
            \item{swap dealers: 'long', 'short', 'net', 'spreading'.}
            \item{other reportables: 'long', 'short', 'net', 'spreading'.}
            \item{commercial - non-CIT: 'long', 'short', 'net'.}
            \item{non-commercial - non-CIT: 'long', 'short', 'net', 'spreading'.}
            \item{index traders - non-CIT: 'long', 'short'.}
            \item{index traders: 'long', 'short', 'net'.}
            \item{asset manager/institutional: 'long', 'short', 'net', 'spreading'.}
            \item{dealer/intermediary: 'long', 'short', 'net', 'spreading'.}
            \item{leveraged funds: 'long', 'short', 'net', 'spreading'.}
            \item{other reportables: 'long', 'short', 'net', 'spreading'.}
         }
       }
     }
   }
    \item{\code{data}: a tibble. Columns inlude:
      \itemize{
        \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
        \item{\code{format}: report formats for which data has been found.}
        \item{\code{underlying}: underlying instruments for which data has been found..}
        \item{\code{unit}: counting units for which data has been found.}
        \item{\code{participant}: trader classifications for which data has been found.}
        \item{\code{position}: trader positions for which data has been found.}
        \item{\code{date}: observation date.}
        \item{\code{value}: observed value}
      }
    }
 }
}
\description{
Provided with a set of Bloomberg futures active contract tickers and a time period,
  queries Bloomberg for the corresponding futures CFTC reports data.
}
\examples{
\dontrun{bbg_futures_CFTC()}

}
\seealso{
The \code{\link[bbgsymbols]{tickers_cftc}} dataset in the \code{bbgsymbols} package for details on the Bloomnerg fields used here.
}
