% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{BBG_pull_historical_books}
\alias{BBG_pull_historical_books}
\title{Pull historical book data from Bloomberg}
\usage{
BBG_pull_historical_books(tickers, field, start, end, ...)
}
\arguments{
\item{tickers}{A character vector. Specifies the Bloomberg tickers for the query}

\item{field}{A character vector. Specifies the Bloomberg field for the query.}

\item{start}{A scalar character vector. Specifies the starting date for the query
in the following format: 'yyyy-mm-dd'.}

\item{end}{A scalar character vector. Specifies the end date for the query in the
following format: 'yyyy-mm-dd'.}

\item{...}{Optional parameters to pass to the \link[Rblpapi]{bdp} function from the
\href{http://dirk.eddelbuettel.com/code/rblpapi.html}{\pkg{Rblpapi}} package used
  for the query (\code{options} parameter).}
}
\description{
Pull historical book data from Bloomberg
}
