% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_call}
\alias{get_call}
\alias{get_call,DataInfo-method}
\alias{get_call,DataInfo}
\alias{get_call,DataHistorical-method}
\alias{get_call,DataHistorical}
\title{Accesses the \code{call} slot of various S4 objects in the
  \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite.}
\usage{
get_call(object)

\S4method{get_call}{DataInfo}(object)

\S4method{get_call}{DataHistorical}(object)
}
\arguments{
\item{object}{an S4 object of classes \linkS4class{DataInfo} or
\linkS4class{DataHistorical} or respective childs.}
}
\value{
The original call to the constructor function.
}
\description{
Access method for the \code{call} slot of various S4
  objects in the
  \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite.
}
