% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{get_data}
\alias{get_data}
\alias{get_data,DataHistorical-method}
\alias{get_data,DataHistorical}
\title{Accesses the \code{data} slot of various S4 objects from
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}.}
\usage{
get_data(object)

\S4method{get_data}{DataHistorical}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataHistorical} or childs.}
}
\value{
A \link[data.table]{data.table} showing for each Bloomberg ticker and
  Bloomberg field the data contained in the \code{object} provided.
}
\description{
Access method for the \code{data} slot of various S4 objects from the
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.
}
