% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.R
\docType{methods}
\name{get_periods}
\alias{get_periods}
\alias{get_periods,FuturesTS-method}
\alias{get_periods,BBGHistorical,FuturesTS}
\alias{get_periods,FuturesAggregate-method}
\alias{get_periods,BBGHistorical,FuturesAggregate}
\alias{get_periods,FuturesCFTC-method}
\alias{get_periods,BBGHistorical,FuturesCFTC}
\alias{get_periods,EquityMarket-method}
\alias{get_periods,BBGHistorical,EquityMarket}
\alias{get_periods,EquityKS-method}
\alias{get_periods,BBGHistorical,EquityKS}
\alias{get_periods,EquityBS-method}
\alias{get_periods,BBGHistorical,EquityBS}
\alias{get_periods,EquityIS-method}
\alias{get_periods,BBGHistorical,EquityIS}
\alias{get_periods,EquityCF-method}
\alias{get_periods,BBGHistorical,EquityCF}
\alias{get_periods,EquityRatios-method}
\alias{get_periods,BBGHistorical,EquityRatios}
\title{Generic method for accessing the 'periods' for which various S4 objects from the \code{pullit} package contain data.}
\usage{
get_periods(object)

\S4method{get_periods}{FuturesTS}(object)

\S4method{get_periods}{FuturesAggregate}(object)

\S4method{get_periods}{FuturesCFTC}(object)

\S4method{get_periods}{EquityMarket}(object)

\S4method{get_periods}{EquityKS}(object)

\S4method{get_periods}{EquityBS}(object)

\S4method{get_periods}{EquityIS}(object)

\S4method{get_periods}{EquityCF}(object)

\S4method{get_periods}{EquityRatios}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{BBGHistorical} or childs.}

\item{object}{an S4 object of class \linkS4class{FuturesTS}.}

\item{object}{an S4 object of class \linkS4class{FuturesAggregate}.}

\item{object}{an S4 object of class \linkS4class{FuturesCFTC}.}

\item{object}{an S4 object of class \linkS4class{EquityMarket}.}

\item{object}{an S4 object of class \linkS4class{EquityKS}.}

\item{object}{an S4 object of class \linkS4class{EquityBS}.}

\item{object}{an S4 object of class \linkS4class{EquityIS}.}

\item{object}{an S4 object of class \linkS4class{EquityCF}.}

\item{object}{an S4 object of class \linkS4class{EquityRatios}.}
}
\value{
A tibble.

For S4 objects of class \linkS4class{FuturesTS} columns include:
   \itemize{
     \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
     \item{\code{TS position}: futures term structure positions for which data has been found.}
     \item{\code{field}: futures market data fields for which data has been found.}
     \item{\code{start}: oldest date with observed data.}
     \item{\code{end}: most recent date with observed data.}
   }

For S4 objects of class \linkS4class{FuturesAggregate} columns include:
   \itemize{
     \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
     \item{\code{field}: futures Bloomberg aggregate data fields for which data has been found.}
     \item{\code{start}: oldest date with observed data.}
     \item{\code{end}: most recent date observed with data.}
   }

For S4 objects of class \linkS4class{FuturesCFTC} columns include:
  \itemize{
    \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
    \item{\code{format}: report formats for which data has been found.}
    \item{\code{underlying}: underlying instruments for which data has been found..}
    \item{\code{unit}: counting units for which data has been found.}
    \item{\code{participant}: trader classifications for which data has been found.}
    \item{\code{position}: trader positions for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date observed with data.}
  }

For S4 objects of class \linkS4class{EquityMarket} columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{field}: equity Bloomberg market data fields for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }

For S4 objects of class \linkS4class{EquityKS} columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{section}: balance sheet sections for which data has been found.}
    \item{\code{subsection}: balance sheet subsections for which data has been found.}
    \item{\code{name}: balance sheet Bloomberg data field names for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }

For S4 objects of class \linkS4class{EquityBS} columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{section}: balance sheet sections for which data has been found.}
    \item{\code{subsection}: balance sheet subsections for which data has been found.}
    \item{\code{name}: balance sheet Bloomberg data field names for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }

For S4 objects of class \linkS4class{EquityIS} columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{name}: income statement Bloomberg field names for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }

For S4 objects of class \linkS4class{EquityCF} columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{section}: cash flow statement sections for which data has been found.}
    \item{\code{name}: cash flow statement Bloomberg field names for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }

For S4 objects of class \linkS4class{EquityRatios} columns include:
  \itemize{
    \item{\code{ticker}: equity Bloomberg tickers for which data has been found.}
    \item{\code{type}: financial ratio types for which data has been found.}
    \item{\code{name}: financial ratio Bloomberg data field names for which data has been found.}
    \item{\code{start}: oldest date with observed data.}
    \item{\code{end}: most recent date with observed data.}
  }
}
\description{
Generic method for accessing the 'periods' for which various S4 objects from the \code{pullit} package contain data.

'periods' method for S4 objects of class \linkS4class{FuturesTS}.

'periods' method for S4 objects of class \linkS4class{FuturesAggregate}.

'periods' method for S4 objects of class \linkS4class{FuturesCFTC}.

'periods' method for S4 objects of class \linkS4class{EquityMarket}.

'periods' method for S4 objects of class \linkS4class{EquityKS}.

'periods' method for S4 objects of class \linkS4class{EquityBS}.

'periods' method for S4 objects of class \linkS4class{EquityIS}.

'periods' method for S4 objects of class \linkS4class{EquityCF}.

'periods' method for S4 objects of class \linkS4class{EquityRatios}.
}
