% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{storethat_futures_cftc}
\alias{storethat_futures_cftc}
\title{Futures CFTC position historical data from Bloomberg}
\usage{
storethat_futures_cftc(file = NULL,
  active_contract_tickers = "C A Comdty", start = "2018-01-01",
  end = "2018-06-30", verbose = TRUE, ...)
}
\arguments{
\item{file}{A scalar chatacter vector. Specifies the target \code{storethat} SQLite database file.}

\item{active_contract_tickers}{A chatacter vector. Specifies the Bloomberg futures active contract tickers to query data for.
Defaults to 'C A Comdty', the Bloomberg active contract ticker for the XCBT corn futures series.}

\item{start}{A scalar character vector. Specifies the starting date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-01-01'.}

\item{end}{A scalar character vector. Specifies the end date for the query in the following format: 'yyyy-mm-dd'.
Defaults to '2018-06-30'.}

\item{verbose}{A logical scalar vector. Should progression messages be printed? Defaults to TRUE.}

\item{...}{optional extra parameters for grouping:
\itemize{
  \item{\code{asset_class}: a character vector. See column 'asset class' in \code{\link[bbgsymbols]{tickers_futures}}.}
  \item{\code{sector}: a character vector. See column 'sector' in \code{\link[bbgsymbols]{tickers_futures}}.}
  \item{\code{subsector}: a character vector. See column 'subsector' in \code{\link[bbgsymbols]{tickers_futures}}.}
  \item{\code{exchange}: a character vector. See column 'symbol' in \code{\link[fewISOs]{exchanges}}.}
  \item{\code{currency}: a character vector. See column 'symbol' in \code{\link[fewISOs]{currencies}}.}
}}
}
\value{
An S4 object of class \code{\linkS4class{FuturesCFTC}} with slots:
  \itemize{
    \item{\code{tickers}: a data.table. Active contract and corresponding position Bloomberg tickers for which data has been found. Columns include:
      \itemize{
        \item{active contract ticker: futures series active contract Bloomberg ticker.}
        \item{position ticker: CFTC position Bloomberg ticker.}
      }
    }
    \item{\code{fields}: a data.table. Columns include:
      \itemize{
        \item{\code{format}: report formats ('legacy', 'disaggregated', 'supplemental' or 'traders in financial futures')
          for which data has been found.}
        \item{\code{underlying}: underlying instruments ('futures only' or 'futures & options') for which data has been found..}
        \item{\code{unit}: counting units (number of 'contracts', 'traders' or 'total') for which data has been found.}
        \item{\code{participant}: trader classifications for which data has been found. Report specific:
          \itemize{
            \item{legacy: 'commercial', 'non-commercial', 'non-reportable', 'total'.}
            \item{disaggregated: 'managed money', 'producer/merchant/processor/user', 'swap dealers', 'other reportables'.}
            \item{supplemental: 'commercial - non-CIT', 'non-commercial - non-CIT', 'index traders - non-CIT', 'index traders'.}
            \item{traders in financial futures: 'asset manager/institutional', 'dealer/intermediary', 'leveraged funds', 'other reportables'.}
          }
        }
        \item{\code{position}: trader positions for which data has been found. Participant specific.
          \itemize{
            \item{commercial: 'long', 'short', 'net'.}
            \item{non-commercial: 'long', 'short', 'net', 'spreading'.}
            \item{non-reportable: 'long', 'short', 'net'.}
            \item{total: 'long', 'short', 'net', 'open interest', 'total'.}
            \item{managed money: 'long', 'short', 'net', 'spreading'.}
            \item{producer/merchant/processor/user: 'long', 'short', 'net'.}
            \item{swap dealers: 'long', 'short', 'net', 'spreading'.}
            \item{other reportables: 'long', 'short', 'net', 'spreading'.}
            \item{commercial - non-CIT: 'long', 'short', 'net'.}
            \item{non-commercial - non-CIT: 'long', 'short', 'net', 'spreading'.}
            \item{index traders - non-CIT: 'long', 'short'.}
            \item{index traders: 'long', 'short', 'net'.}
            \item{asset manager/institutional: 'long', 'short', 'net', 'spreading'.}
            \item{dealer/intermediary: 'long', 'short', 'net', 'spreading'.}
            \item{leveraged funds: 'long', 'short', 'net', 'spreading'.}
            \item{other reportables: 'long', 'short', 'net', 'spreading'.}
         }
       }
       \item{\code{position ticker}: corresponding CFTC position Bloomberg ticker.}
     }
   }
    \item{\code{data}: a tibble. Columns inlude:
      \itemize{
        \item{\code{active contract ticker}: futures active contract Bloomberg tickers for which data has been found.}
        \item{\code{position ticker}: CFTC position Bloomberg tickers for which data has been found.}
        \item{\code{date}: observation date.}
        \item{\code{value}: observed value}
      }
    }
    \item{\code{call}: a scalar character vector showing the original call to the constructor.}
 }
}
\description{
Provided with a set of Bloomberg futures active contract tickers and a time period,
  retrieves the corresponding futures CFTC position historical data previously stored in a \code{storethat}
  SQLite database.
}
\examples{
\dontrun{storethat_futures_cftc()}

}
\seealso{
The \code{\link[bbgsymbols]{tickers_cftc}} dataset in the \code{bbgsymbols} package for details on the Bloomnerg fields used here.
}
